feat(appeals): add pdf file to bulk file download and add pdf service url to config
