+++ appeals/pdf-service-api/.env.example
@@ -1,0 +1,5 @@
+AUTH_BASE_URL=
+LOGGER_LEVEL=info
+SERVER_PORT=3000
+SERVER_SHOW_ERRORS=true
+SERVER_TERMINATION_GRACE_PERIOD_SECONDS=60
