+++ appeals/pdf-service-api/.gitignore
@@ -1,1 +1,2 @@
 *.pdf
+/src/public/assets/*.css
