+++ appeals/pdf-service-api/src/config.js
@@ -1,17 +1,24 @@
-const config = {
+import { loadEnvironment } from '@pins/platform';
+import path from 'node:path';
+import url from 'node:url';
+
+const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
+const apiDir = path.join(__dirname, '..'); // package root, where .env files live
+
+const environment = loadEnvironment(process.env.NODE_ENV, apiDir);
+
+export default {
 	gitSha: process.env.GIT_SHA ?? 'NO GIT SHA FOUND',
 	auth: {
-		authServerUrl: process.env.AUTH_BASE_URL
+		authServerUrl: environment.AUTH_BASE_URL
 	},
 	logger: {
-		level: process.env.LOGGER_LEVEL || 'info'
+		level: environment.LOGGER_LEVEL || 'info'
 	},
 	server: {
-		port: Number(process.env.SERVER_PORT) || 3000,
-		showErrors: process.env.SERVER_SHOW_ERRORS === 'true',
+		port: Number(environment.SERVER_PORT) || 3000,
+		showErrors: environment.SERVER_SHOW_ERRORS === 'true',
 		terminationGracePeriod:
-			(Number(process.env.SERVER_TERMINATION_GRACE_PERIOD_SECONDS) || 0) * 1000
+			(Number(environment.SERVER_TERMINATION_GRACE_PERIOD_SECONDS) || 0) * 1000
 	}
 };
-
-module.exports = config;
