+++ appeals/pdf-service-api/src/lib/logger.js
@@ -1,8 +1,8 @@
-const { pino } = require('pino');
-const config = require('../config');
+import { pino } from 'pino';
+import config from '../config.js';
 
 const logger = pino({
 	level: config.logger.level
 });
 
-module.exports = logger;
+export default logger;
