+++ appeals/web/.env.example
@@ -44,3 +44,5 @@ USE_SYSTEM_TEST_BC_FOR_CHANGE_LPA=false
 SESSION_SECRET=SomeSecretHere
 
 BETA_FEEDBACK_URL=https://forms.office.com/Pages/ResponsePage.aspx?id=mN94WIhvq0iTIpmM5VcIjYt1ax_BPvtOqhVjfvzyJN5UQzNVQjdaV0U0TDdTOE1LNktBT0w5NEQ1Vy4u
+
+PDF_SERVICE_HOST=http://localhost:3000
