+++ appeals/web/src/server/appeals/appeal-details/lpa-questionnaire/lpa-questionnaire.service.js
@@ -3,7 +3,7 @@
  */
 
 /**
- * @param {string} appealId
+ * @param {string|number} appealId
  * @param {string} lpaQuestionnaireId
  * @param {import('got').Got} apiClient
  * @returns {Promise<LpaQuestionnaire>}
